backend/requirements.txt
fastapi
uvicorn
transformers
torch
openai
pandas
python-dotenv
aiohttp
